生成了论坛相关，实现了多文件上传，文件上传服务器测试且与前端对接成功，实现了帖子信息表的查询

+++ src/main/java/pers/lcf/rents/userbase/model/UserMatch.java
@@ -15,6 +15,6 @@ public class UserMatch {
     private String avatar;
     private String sex;
     private String statement;
-    private Integer similarity;
+    private Float similarity;
     private UserStyle userStyle;
 }
